Make `TagInput` a controlled component. Fixes #11
Major breaking changes:
- `TagInput` is a controlled component and does not maintain tags in its state
- `transformTag`, `onChange` and `getTagStyle` now work on tag objects, not strings

This isn't what we discussed in discussion of #11. I found
in implementation that letting the component work as both controlled and uncontrolled introduced
a ton of complexity and really hurt the readability and maintainability of the code.

I also discovered that `transformTag` is really useful: without it, the user of this library would
have to implement encoding and decoding of tag metadata (including category) to/from strings,
which seems like too much effort for something that should be trivial.

Using an array of objects for tags rather than an array of strings, together with `transformTag`,
provides a simple and flexible way to track the state and control what text the user sees. The objects
also allow adding arbitrary data to tags, which could be useful for things like controlling the tag style.

Looking forward to your feedback